strjoin

+++ ft_atoi.c
@@ -17,9 +17,9 @@ static int	ft_isspace(int c);
 /* ************************************************************************** */
 int	ft_atoi(const char *str)
 {
-	size_t		i;
-	long long	sign;
-	long long	value;
+	size_t	i;
+	int		sign;
+	int		value;
 
 	sign = 1;
 	i = 0;
@@ -34,11 +34,7 @@ int	ft_atoi(const char *str)
 	value = 0;
 	while ((str[i] != '\0') & ft_isdigit(str[i]))
 	{
-		// printf("index: %zu | long long value %lld | value: %d\n", i, value, (int)value);
-		// sk_printmem(&value, 24);
 		value = value * 10 + (str[i] - 48);
-		// if ((value > 922337203685477579) & (str[i + 1] != '\0') & ft_isdigit(str[i + 1]))
-		// 	return (-1);
 		i++;
 	}
 	value *= sign;
